Consolidate common shared to_formatted_s method
Core extensions for Date and Time both included identical code
for to_formatted_s. So, following convention established by
other date_and_time shared methods, I extracted it out into
a common module.
